Engine refactor. Interface:engine. Classes engine.model
