+++ engine/model/capulet_engine.py
@@ -1,0 +1,25 @@
+from abc import ABC
+
+from engine import engine
+
+from datetime import datetime
+
+
+
+class CapuletEngine(engine, ABC):
+    def __init__(self, last_service_date, current_mileage, last_service_mileage):
+        super().__init__(last_service_date)
+        self.current_mileage = current_mileage
+        self.last_service_mileage = last_service_mileage
+
+        
+    def engine_should_be_serviced(self):
+        return self.current_mileage - self.last_service_mileage > 30000
+
+
+    def needs_service(self):
+        service_threshold_date = self.last_service_date.replace(year=self.last_service_date.year + 2)
+        if service_threshold_date < datetime.today().date() or self.engine_should_be_serviced():
+            return True
+        else:
+            return False
